API - InLock

+++ SPRINT-2-API/Application Programming Interface (API)/API-InLock/senai.inlock.webApi/Domain/JogoDomain.cs
@@ -20,7 +20,7 @@ namespace senai.inlock.webApi.Domain
 
 
         [Required(ErrorMessage = "Valor do jogo obrigatório !")]
-        public float Valor { get; set; }
+        public Decimal Valor { get; set; }
 
 
         //instância de EstudioDomain
